Translation update !

+++ src/Secret/Application/CommandHandler/DeleteSecretHandler.php
@@ -4,6 +4,7 @@ namespace Marvin\Secret\Application\CommandHandler;
 
 use EnderLab\DddCqrsBundle\Application\Command\SyncCommandHandlerInterface;
 use Marvin\Secret\Application\Command\DeleteSecret;
+use Marvin\Secret\Domain\Exception\SecretNotFound;
 use Marvin\Secret\Domain\Repository\SecretRepositoryInterface;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 
@@ -18,6 +19,11 @@ final readonly class DeleteSecretHandler implements SyncCommandHandlerInterface
     public function __invoke(DeleteSecret $command): void
     {
         $secret = $this->secretRepository->byKey($command->key);
+
+        if (null === $secret) {
+            throw SecretNotFound::withKey($command->key);
+        }
+
         $secret->delete();
 
         $this->secretRepository->remove($secret);
